requirements.txt
flask
mysql-connector-python
opencv-python-headless
deepface
numpy
Pillow
gunicorn
tf-keras